requirements.txt
requests==2.20.0
lxml==4.3.1
prettytable==0.7.2
Pillow==5.4.1
beautifulsoup4==4.7.1
